refactor: code style
